docker/emailservice/requirements.txt
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
elastic-apm==5.10.0
google-api-core==1.6.0
google-api-python-client==1.7.8
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-cloud-core==0.29.1
googleapis-common-protos==1.5.5
grpcio==1.24.0
grpcio-health-checking==1.13.0
grpcio-tools==1.24.0
httplib2==0.12.1
idna==2.8
Jinja2==2.10
MarkupSafe==1.1.0
protobuf==3.14.0
psutil==5.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.3
python-json-logger==0.1.9
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
ecs-logging==0.2
structlog==20.1.0